adds split in fast and hifi resampler module
