+++ src/conversions/sample_rate/hifi_rubato.rs
@@ -1,0 +1,60 @@
+use crate::Sample;
+
+#[derive(Clone, Debug)]
+pub struct SampleRateConverter<I>
+where
+    I: Iterator,
+{
+    /// The iterator that gives us samples.
+    input: I,
+    /// We convert chunks of `from` samples into chunks of `to` samples.
+    from: u32,
+    /// We convert chunks of `from` samples into chunks of `to` samples.
+    to: u32,
+    /// Number of channels in the stream
+    channels: cpal::ChannelCount,
+    /// One sample per channel, extracted from `input`.
+    current_frame: Vec<I::Item>,
+    /// Position of `current_sample` modulo `from`.
+    current_frame_pos_in_chunk: u32,
+    /// The samples right after `current_sample` (one per channel), extracted from `input`.
+    next_frame: Vec<I::Item>,
+    /// The position of the next sample that the iterator should return, modulo `to`.
+    /// This counter is incremented (modulo `to`) every time the iterator is called.
+    next_output_frame_pos_in_chunk: u32,
+    /// The buffer containing the samples waiting to be output.
+    output_buffer: Vec<I::Item>,
+}
+
+impl<I> SampleRateConverter<I>
+where
+    I: Iterator,
+    I::Item: Sample,
+{
+    #[inline]
+    pub fn new(
+        mut input: I,
+        from: cpal::SampleRate,
+        to: cpal::SampleRate,
+        num_channels: cpal::ChannelCount,
+    ) -> SampleRateConverter<I> {
+        let from = from.0;
+        let to = to.0;
+
+        assert!(num_channels >= 1);
+        assert!(from >= 1);
+        assert!(to >= 1);
+
+        SampleRateConverter {
+            input,
+            from,
+            to,
+            channels: num_channels,
+            current_frame_pos_in_chunk: 0,
+            next_output_frame_pos_in_chunk: 0,
+            current_frame: first_samples,
+            next_frame: next_samples,
+            output_buffer: Vec::with_capacity(num_channels as usize - 1),
+        }
+    }
+}
